Fix linting errors that are only caught by stylguidist for some reason
